+++ udata_front/theme/gouvfr/datagouv-components/src/helpers/index.ts
@@ -75,7 +75,7 @@ export const formatFromNow = (date: Date | string) => {
   if(today.isSame(dateWithoutTime)) {
     return t("today");
   }
-  return dayjs(date).fromNow();
+  return dateWithoutTime.fromNow();
 }
 
 /**